requirements.txt
streamlit==1.44.1
requests==2.32.3
google-generativeai==0.8.5
firecrawl-py==1.16.0